+++ src/features/common/shared/auth/modules/Header/__tests__/propertyFromHeader.test.ts
